scripts_ai/requirements.txt
PyGithub==1.59.0
GitPython==3.1.32
requests

